move the vector relays into their own translation unit
server_state.c reached 2510 lines and line-check caps it at 2500. Same reason
server_bearer_auth.c exists: lift the new code out rather than push an existing
file further over.

server_state_vectors.c holds kb.reembed, memory.embed and the checked relay they
share.
